// Client/src/api/ServerApi.ts
export interface UserDto {
    userId?: string;
    username?: string;
    discordUsername?: string | undefined;
    email?: string;
    role?: string;
}

export interface UpdateUserDto {
    username?: string | undefined;
    discordUsername?: string | undefined;
    email?: string | undefined;
    password?: string | undefined;
    role?: string | undefined;
}